MAINT Move DistanceMetric under metrics
