+++ sklearn/metrics/setup.py
@@ -1,4 +1,5 @@
 import os
+import numpy as np
 
 from numpy.distutils.misc_util import Configuration
 
@@ -16,6 +17,13 @@ def configuration(parent_package="", top_path=None):
 
     config.add_extension(
         "_pairwise_fast", sources=["_pairwise_fast.pyx"], libraries=libraries
+    )
+
+    config.add_extension(
+        "_dist_metrics",
+        sources=["_dist_metrics.pyx"],
+        include_dirs=[np.get_include(), os.path.join(np.get_include(), "numpy")],
+        libraries=libraries,
     )
 
     config.add_subpackage("tests")
