+++ sklearn/neighbors/setup.py
@@ -33,19 +33,6 @@ def configuration(parent_package="", top_path=None):
     )
 
     config.add_extension(
-        "_dist_metrics",
-        sources=["_dist_metrics.pyx"],
-        include_dirs=[numpy.get_include(), os.path.join(numpy.get_include(), "numpy")],
-        libraries=libraries,
-    )
-
-    config.add_extension(
-        "_typedefs",
-        sources=["_typedefs.pyx"],
-        include_dirs=[numpy.get_include()],
-        libraries=libraries,
-    )
-    config.add_extension(
         "_quad_tree",
         sources=["_quad_tree.pyx"],
         include_dirs=[numpy.get_include()],
