+++ sklearn/utils/setup.py
@@ -88,6 +88,13 @@ def configuration(parent_package="", top_path=None):
         libraries=libraries,
     )
 
+    config.add_extension(
+        "_typedefs",
+        sources=["_typedefs.pyx"],
+        include_dirs=[numpy.get_include()],
+        libraries=libraries,
+    )
+
     config.add_subpackage("tests")
 
     return config
